[+-] Moved Queue to it's own module
This is to keep the main file as clean as possible, and for ease of
access for making any future changes to queues.
