+++ TCGServer.py
@@ -1,6 +1,6 @@
 import sys, time, socketserver
 import emailinfo, regrules, TCGMain
-from modules import pyrand, pyemail, pyhash
+from modules import pyrand, pyemail, pyhash, pyqueue
 from threading import Thread
 from os import walk
 
@@ -11,36 +11,6 @@ class SimpleServer(socketserver.ThreadingMixIn, socketserver.TCPServer):
 
     def __init__(self, server_address, request_handler):
         socketserver.TCPServer.__init__(self, server_address, request_handler)
-
-
-class Queue:
-    def __init__(self, qtype='f', maxsize=0):
-        if qtype.lower()[0] in 'lf':
-            self.qtype = qtype
-        else:
-            raise TypeError('qtype must be (l) or (f)')
-        if maxsize < 0:
-            self.maxsize = 0
-        else:
-            self.maxsize = maxsize
-        self.queue = []
-
-    def put(self, item):
-        if self.maxsize > 0:
-            if len(self.queue) == self.maxsize:
-                raise ValueError('Queue is full')
-        self.queue.append(item)
-
-    def get(self):
-        if self.qtype == 'f':
-            return self.queue.pop(0)
-        elif self.qtype == 'l':
-            return self.queue.pop(len(self.queue)-1)
-        else:
-            raise TypeError('Type must be (l) or (f)')
-
-    def empty(self):
-        return not bool(self.queue)
 
 
 class UserHandler(socketserver.BaseRequestHandler):
@@ -287,4 +257,7 @@ if __name__ == "__main__":
             workers[queue].start()
         server.serve_forever()
     except KeyboardInterrupt:
+        sys.exit(0)
+    except Exception as e:
+        print(e)
         sys.exit(0)
